third_party/WebKit/LayoutTests/http/tests/inspector/network/network-fetch-expected.txt

Tests fetch network resource type and content.

http://127.0.0.1:8000/inspector/network/resources/resource.php
resource.type: fetch
resource.content before requesting content: undefined
resource.content after requesting content: Hello world
